@@ -33,6 +33,15 @@ import {FormsModule} from '@angular/forms';
 import { HomeComponent } from './ui/home/home.component';
 import { FacebookProfilePictureComponent } from './users/facebook-profile-picture/facebook-profile-picture.component';
 import { SolutionComponent } from './ui/solution/solution.component';
+import { AdminComponent } from './ui/admin/admin.component';
+import { FacultyControlPanelComponent } from './ui/admin/faculty-control-panel/faculty-control-panel.component';
+import { TopicComponent } from './ui/topic/topic.component';
+import { CommentComponent } from './ui/comment/comment.component';
+import { CreateTopicComponent } from './ui/create-topic/create-topic.component';
+import {CKEditorModule} from 'ngx-ckeditor';
+import { CreateCommentComponent } from './ui/create-comment/create-comment.component';
+import { FacultiesComponent } from './ui/faculties/faculties.component';
+import { FacultyComponent } from './ui/faculty/faculty.component';
 
 @NgModule({
   declarations: [
@@ -50,7 +59,15 @@ import { SolutionComponent } from './ui/solution/solution.component';
     FacebookProfilePictureComponent,
     ScanPageComponent,
     UploadQuestionComponent,
-    SolutionComponent
+    SolutionComponent,
+    AdminComponent,
+    FacultyControlPanelComponent,
+    TopicComponent,
+    CommentComponent,
+    CreateTopicComponent,
+    CreateCommentComponent,
+    FacultiesComponent,
+    FacultyComponent
   ],
   imports: [
     BrowserModule,
@@ -66,7 +83,8 @@ import { SolutionComponent } from './ui/solution/solution.component';
     ImageCropperModule,
     CoreModule,
     MatSnackBarModule,
-    FormsModule
+    FormsModule,
+    CKEditorModule
   ],
   providers: [AuthService, PdfService],
   bootstrap: [AppComponent]
