@@ -3,6 +3,8 @@ import {ActivatedRoute} from '@angular/router';
 import {DbService} from '../../core/db.service';
 import {Question, QuestionId} from '../../core/entities/question';
 import {Exam} from '../../core/entities/exam';
+import {AuthService} from '../../core/auth.service';
+import {Course} from '../../core/entities/course';
 
 @Component({
   selector: 'app-exam',
@@ -15,10 +17,15 @@ export class ExamComponent implements OnInit {
   private examId: string;
   public questions: QuestionId[];
   public exam: Exam;
+  course: Course;
 
-  constructor(private route: ActivatedRoute, private db: DbService) {
+  adminAccess: boolean;
+
+  constructor(private route: ActivatedRoute, private db: DbService, private auth: AuthService) {
     this.courseId = +route.snapshot.paramMap.get('cid');
     this.examId = route.snapshot.paramMap.get('eid');
+    this.auth.isAdminForCourse(this.courseId).subscribe(is => this.adminAccess = is);
+    this.db.getCourse(this.courseId).subscribe(course => this.course = course);
   }
 
   ngOnInit() {
