@@ -11,12 +11,14 @@ export class HeaderComponent implements OnInit {
   main_menu_opened: boolean;
   @Output() main_menu_triggered: EventEmitter<boolean>;
   @Input() term: any;
-  router: any
+  router: any;
+  isAdmin: boolean;
 
   constructor(private rtr: Router, public auth: AuthService) {
     this.main_menu_opened = false;
     this.main_menu_triggered = new EventEmitter();
     this.router = rtr;
+    this.auth.isAdmin.subscribe(res => this.isAdmin = res);
   }
 
   trigger_menu() {
