@@ -5,12 +5,12 @@
 export const environment = {
   production: false,
   firebase: {
-    apiKey: 'AIzaSyCREv45Llml_gg4hwXeMTXrB6gsDdBkSUQ',
-    authDomain: 'yp5779-testock.firebaseapp.com',
-    databaseURL: 'https://yp5779-testock.firebaseio.com',
-    projectId: 'yp5779-testock',
-    storageBucket: 'yp5779-testock.appspot.com',
-    messagingSenderId: '997206110546'
+    apiKey: 'AIzaSyAqyOrcrmIdg2yMylYkvnKLpNk8AWqhWuk',
+    authDomain: 'testock-development.firebaseapp.com',
+    databaseURL: 'https://testock-development.firebaseio.com',
+    projectId: 'testock-development',
+    storageBucket: 'testock-development.appspot.com',
+    messagingSenderId: '398800867598'
   }
 };
 
